% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTag.R
\name{extractTag}
\alias{extractTag}
\title{Remove a tag from an HTML file}
\usage{
extractTag(filecon, tag)
}
\arguments{
\item{filecon}{File connection}

\item{tag}{Tag name}
}
\description{
Remove a tag from an HTML file
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

