% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minifyHtml.R
\name{minifyHtml}
\alias{minifyHtml}
\title{Minify an HTML file into a single string removing unnecesary spaces}
\usage{
minifyHtml(input, output = "output.html", rm.comments = FALSE,
  wb.remove = TRUE)
}
\arguments{
\item{input}{HTML input file.}

\item{output}{Name for the output file.}

\item{rm.comments}{Remove comments?}

\item{wb.remove}{Remove Wayback Machine comments?}
}
\description{
Minify an HTML file into a single string removing unnecesary spaces
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

