% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parNesting.R
\name{parNesting}
\alias{parNesting}
\title{Newick representation of an HTML tree}
\usage{
parNesting(a)
}
\arguments{
\item{a}{Representation of the tree as a linear vector with nodes numbers (nested set model).}
}
\description{
Newick representation of a tree
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

