% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTag.R
\name{removeTag}
\alias{removeTag}
\title{Remove a tag from an HTML file}
\usage{
removeTag(filecon, tag, single = FALSE)
}
\arguments{
\item{filecon}{HTML input file.}

\item{tag}{Name of the tag.}

\item{single}{Is this a single (not-closed) tag?}
}
\description{
Remove a tag from an HTML file
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

