% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectbyTag.R
\name{selectbyTag}
\alias{selectbyTag}
\title{Filter the SQLite database by tag and attributes}
\usage{
selectbyTag(my_db, tag, attrList = NULL, out_db, chunkSize = 500)
}
\arguments{
\item{my_db}{Input SQLite database.}

\item{tag}{Tag used to filter the database.}

\item{attrList}{List of attributes of the selected tag, used to filter the database.}

\item{out_db}{Name of the output (filtered) database.}

\item{chunkSize}{Number of chunks used to write the output database.}
}
\description{
Filter the SQLite database by tag and attributes
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

