% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendQuery.R
\name{sendQuery}
\alias{sendQuery}
\title{Send an SQL query to SQLite database}
\usage{
sendQuery(myDb, myQuery)
}
\arguments{
\item{myDb}{SQLite database}

\item{myQuery}{SQLite query}
}
\description{
Send an SQL query to SQLite database
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

