% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveTree.R
\name{solveTree}
\alias{solveTree}
\title{Find nodes and branchs of an HTML tree in a nested set model representation.}
\usage{
solveTree(a, recursive = TRUE)
}
\arguments{
\item{a}{Representation of the tree as a linear vector nodes  numbers (nested set model)}
}
\description{
Find nodes and branchs of an HTML tree in a nested set model representation
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

