% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subUnquoted.R
\name{subUnquoted}
\alias{subUnquoted}
\title{Change text as a regex pattern, inside or outside quotes.}
\usage{
subUnquoted(x, from, to, invert = TRUE, perl = TRUE)
}
\arguments{
\item{x}{Input text}

\item{invert}{If TRUE, changes are made over text within quotes.}

\item{From}{pattern to substitute}

\item{To}{substitution}
}
\description{
Change text as a regex pattern, inside or outside quotes
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

