% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagExplitit.R
\name{tagExplicit}
\alias{tagExplicit}
\title{Reconstruct an implicit HTML expression of the SQL database}
\usage{
tagExplicit(my_db, tagId, maxLevel = 1)
}
\arguments{
\item{my_db}{Database.}

\item{tagId}{number of tag.}

\item{maxLevel}{Maximum level of nested tags converted into explicit expressions.
Default 1, for all the levels explicit use te value "ALL".}
}
\description{
reconstruct an implicit HTML expression of the SQL database
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

