% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waybackRemove.R
\name{waybackRemove}
\alias{waybackRemove}
\title{Remove Wayback Machine content included in a page}
\usage{
waybackRemove(my_file)
}
\arguments{
\item{my_file}{HTML file to edit.}
}
\description{
Remove Wayback Machine content included in a page
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

