% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waybackScrapper.R
\name{waybackScrapper}
\alias{waybackScrapper}
\title{Obtain a page over a series of dates using the Wayback Machine}
\usage{
waybackScrapper(myUrl, start, end, by = "hour", intervalVector = NULL)
}
\arguments{
\item{myUrl}{URL of the page}

\item{start}{start date as ISOdatetime object}

\item{end}{end date as as ISOdatetime object}

\item{by}{argument passed to seq, defining the interval. Default "hour".}

\item{IntervalVector}{Optional vector of dates as a sequence with format "%Y%m%d%H%M%S"}
}
\description{
Obtain a page over a series of dates using the Wayback Machine
}
\author{
Leandro Roser \email{leandroroser@gmail.com}
}

